feat: manager users

+++ reactjs/food-deliveries/backend/src/controllers/users.controller.js
@@ -1,7 +1,32 @@
 import Payment from '../models/payments.model';
 import User from '../models/users.model';
+import { createUserValidate } from '../validates/users.validate';
 
 export const userController = {
+	/* create users */
+	create: async (req, res) => {
+		try {
+			const body = req.body;
+			/* validate */
+			await createUserValidate.validate(body);
+			/* check email */
+			const userExist = await User.findOne({ email: body.email });
+			if (userExist) {
+				return res.status(400).json({ message: 'Email already exists' });
+			}
+			/* create User */
+			const user = await User.create(body);
+			if (!user) {
+				return res.status(400).json({ message: 'Create user failed' });
+			}
+			return res.status(200).json({ user });
+		} catch (error) {
+			if (error.errors) {
+				return res.status(400).json({ message: error.errors });
+			}
+			return res.status(500).json({ message: error.message });
+		}
+	},
 	/* get All */
 	getAll: async (req, res) => {
 		try {
@@ -9,6 +34,7 @@ export const userController = {
 			const options = {
 				page: _page,
 				limit: _limit,
+				populate: [{ path: 'paymentMethodId', select: '-userId' }],
 			};
 			/* nếu có tìm kiếm thì sẽ trả về theo giá trị tìm kiếm */
 			if (q) {
@@ -16,6 +42,7 @@ export const userController = {
 					$or: [
 						{ name: { $regex: q, $options: 'i' } },
 						{ email: { $regex: q, $options: 'i' } },
+						{ phone: { $regex: q, $options: 'i' } },
 					],
 				};
 				const users = await User.paginate(querySearch, options);
@@ -57,6 +84,17 @@ export const userController = {
 			if (!user) {
 				return res.status(404).json({ message: 'User not found' });
 			}
+			/* xóa người payment cũ ra khỏi danh sách */
+			await Payment.findByIdAndUpdate(user.paymentMethodId, {
+				$pull: { userId: id },
+			});
+			/* Thêm mới ngươi dùng vào dánh sách */
+			const paymentMethodId = body.paymentMethodId;
+			if (paymentMethodId) {
+				await Payment.findByIdAndUpdate(paymentMethodId, {
+					$push: { userId: id },
+				});
+			}
 			return res.status(200).json({ user });
 		} catch (error) {
 			return res.status(500).json({ message: error.message });
@@ -66,6 +104,11 @@ export const userController = {
 	delete: async (req, res) => {
 		try {
 			const id = req.params.id;
+			/* compare role admin */
+			const userAdmin = await User.findById({ _id: id });
+			if (userAdmin.role === 'admin') {
+				return res.status(401).json({ message: 'You are not authorized' });
+			}
 			const user = await User.findByIdAndDelete(id).populate('paymentMethodId');
 			if (!user) {
 				return res.status(404).json({ message: 'User not found' });
